+++ tests/testthat/test-category_table.R
@@ -14,6 +14,22 @@ test_that("x", {
   expect_equivalent(x, x2)
 })
 
+test_that("case insensitivity", {
+  marks <- data.frame(ID = 7, K = "Good", E = "good", C = "Bad")
+  cols <- c(K = "Knowledge", E = "Evaluation", C = "Communication")
+  cats <- c("Bad", "OK", "Good")
+  symbol <- "*"
+  x <- category_table(marks, cols, cats, symbol, F)
+  x2 <- data.frame(
+    Criteria = factor(cols),
+    Bad = c("", "", symbol),
+    OK = c("", "", ""),
+    Good = c(symbol, symbol, "")
+  )
+  
+  expect_equivalent(x, x2)
+})
+
 test_that("no cats", {
   # numeric cats
   marks <- data.frame(ID = 7, K = 1, E = 2, C = 3)